# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Halls API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/advertisements": {
            "get": {
                "operationId": "AdvertisementController@index",
                "responses": {
                    "200": {
                        "description": "Display a listing of advertisements."
                    }
                }
            }
        },
        "/api/advertisements/{advertisement}": {
            "get": {
                "operationId": "AdvertisementController@show",
                "parameters": [
                    {
                        "name": "advertisement",
                        "in": "path",
                        "description": "ID of advertisement to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display the specified advertisement.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Advertisement not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Advertisement not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}